refactor(order table): order table <-> table group 리펙토링
