+++ src/test/java/kitchenpos/acceptance/product/ProductAcceptance.java
@@ -69,6 +69,6 @@ public class ProductAcceptance extends AcceptanceTest {
 	}
 
 	public static void 상품_등록_실패됨(ExtractableResponse<Response> response) {
-		assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
 	}
 }
